계산기 1 [STEP 2] 사파리

+++ Calculator/CalculatorTests/LinkdeListTests.swift
@@ -9,16 +9,14 @@ import XCTest
 @testable import Calculator
 
 class LinkdeListTests: XCTestCase {
-    var sut: LinkdeList<Int>!
+    var sut: LinkdeList<Double>!
 
     override func setUpWithError() throws {
-        sut = LinkdeList<Int>()
-
+        sut = LinkdeList<Double>()
     }
 
     override func tearDownWithError() throws {
         sut = nil
-
     }
     
     func test_append_했을때_isEmmpy_return_false() {
@@ -77,7 +75,7 @@ class LinkdeListTests: XCTestCase {
     
     func test_append_data_removeFirst_return_data_Equal_true() {
         // given
-        let testData = 1
+        let testData = 1.0
         sut.append(testData)
         
 
@@ -90,7 +88,7 @@ class LinkdeListTests: XCTestCase {
     
     func test_append_3번_FisrtData_removeFirst_return_data_Equal_true() {
         // given
-        let testData = [1, 2, 3]
+        let testData = [1.0, 2.0, 3.0]
         testData.forEach(sut.append(_:))
         
 
@@ -103,7 +101,7 @@ class LinkdeListTests: XCTestCase {
     
     func test_append_3번_removeFirst_1번_count_return_2() {
         // given
-        let testData = [1, 2, 3]
+        let testData = [1.0, 2.0, 3.0]
         testData.forEach(sut.append(_:))
         sut.removeFirst()
         
@@ -117,7 +115,7 @@ class LinkdeListTests: XCTestCase {
     
     func test_append_3번_removeAll_isEmpty_return_true() {
         // given
-        let testData = [1, 2, 3]
+        let testData = [1.0, 2.0, 3.0]
         testData.forEach(sut.append(_:))
         sut.removeAll()
         
@@ -131,7 +129,7 @@ class LinkdeListTests: XCTestCase {
     
     func test_append_3번_removeAll_count_return_0() {
         // given
-        let testData = [1, 2, 3]
+        let testData = [1.0, 2.0, 3.0]
         testData.forEach(sut.append(_:))
         sut.removeAll()
